build-tools: Move markdown utilities into shared md package
Move markdown utilities into shared md package so that we can use the `md.Walk()`
function in a follow-up commit building exercise question parsing and validation.
